requirements.txt
diffusers
transformers
accelerate
safetensors

torch
torchvision
pillow
numpy

opencv-python
huggingface-hub
ultralytics

openai
requests
beautifulsoup4
tldextract
selenium

scikit-image
tqdm
python-dotenv
replicate
